Push workers to a separate module.
